requirements.txt
flask
flask-cors
llama_parse
openai
requests
pinecone-client
llama-index-llms-openai
llama-index-postprocessor-cohere-rerank
python-dotenv
llama-index-agent-openai
llama-index-embeddings-openai
langchain 
pytesseract 
pdf2image 
pillow
poppler
tesseract
llama-index-vector-stores-pinecone
llama-index-embeddings-huggingface

# Ignore for now
# pip install --upgrade pybind11        # 2.10.1 or higher  (latest as of today: 2.11.1)
# pip install --verbose  'nmslib @ git+https://github.com/nmslib/nmslib.git#egg=nmslib&subdirectory=python_bindings'
# pip install -U langchain-openai
# pip install --upgrade llama-index